requirements.txt
Django
gunicorn
psycopg2
Fabric3
django-environ